% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsePatterns.R
\docType{package}
\name{responsePatterns}
\alias{responsePatterns}
\title{responsePatterns: A package for finding instances of careless responding in data sets}
\description{
Some survey participants tend to respond carelessly which complicates data analysis. This package provides functions that make it easier to explore responses and identify those that may be problematic. Two core functions of this package provide two approaches to the problem of careless responses detection: one based on the auto-correlation approach and one based on a mechanistic approach. Both approaches yield scores that serve as estimates of how problematic the observations potentially are. However, no conclusions should be made without a closer inspection of the problematic responses. Any decision about removing or downweighing an observation should be based on visual inspection of the responses, the specifics of the instrument used to collect the data, familiarity with the whole data set and the context of the data collection process.
}
\section{Auto-correlation approach}{

The \code{\link{rp.acors}} function allows for a probabilistic detection of repetitive patterns a in survey data. This function calculates auto-correlation coefficients for all lags up to the value defined by the max.lag parameter for each observation (respondent). Subsequently, it assigns a percentile value to each observation (respondent) based either on the highest absolute auto-correlation or the sum of absolute auto-correlations.
}

\section{Mechanistic approach}{

The \code{\link{rp.patterns}} function searches for repetitive patterns in the data using an iterative algorithm. Patterns are defined based on the data: if a sequence of values occurs more than once within an observation, it is considered a repetition. The algorithm counts the number of repetitions for each pattern length and then weighs this sum by the length of the pattern (longer patterns are assigned higher weight). The total score for each respondent is determined as the sum of these pattern length-specific scores and is standardized to take a value between 0 and 1.
}

\section{Auxiliary functions}{

The package provides auxiliary functions to summarize the responsePatterns object (\code{\link{rp.summary}}), extract indices (\code{\link{rp.indices}}, \code{\link{rp.hist}}, \code{\link{rp.save2csv}}) and to visually inspect individual responses (\code{\link{rp.plot}}, \code{\link{rp.plots2pdf}}).
}

\references{
Gottfried, J., Jezek, S., & Kralova, M. (2021). \emph{Autocorrelation screening: A potentially efficient method for detecting repetitive response patterns in questionnaire data.} Manuscript submitted for review.
}
