% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_acors.R
\name{rp.acors}
\alias{rp.acors}
\title{Auto-correlation patterns}
\usage{
rp.acors(
  data,
  max.lag = NULL,
  id.var = NULL,
  na.rm = FALSE,
  cor.method = c("pearson", "spearman", "kendall"),
  percentile.method = c("max", "sum"),
  store.data = TRUE
)
}
\arguments{
\item{store.data}{}
}
\value{

}
\description{
Auto-correlation patterns
}
