% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_hist.R
\name{rp.hist}
\alias{rp.hist}
\title{Plots a histogram of the main index}
\usage{
rp.hist(rp.object)
}
\arguments{
\item{rp.object}{A ResponsePatterns object.}
}
\value{
Returns a plot.
}
\description{
This function plots a histogram of the main "suspicion" index. The choice of the index depends on the type and setting of the analysis: it is either the maximum absolute auto-correlation or the sum of absolute auto-correlations if analyzed via the rp.acors function and the total score of analyzed via the rp.patterns function.
}
\examples{
rp <- rp.acors(rp.simdata, id.var="optional_ID")
rp.hist(rp)
}
\seealso{
\code{\link{rp.acors}}, \code{\link{rp.patterns}}
}
