% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_select.R
\name{rp.select}
\alias{rp.select}
\title{Select observations}
\usage{
rp.select(rp.object, percentile = 90)
}
\arguments{
\item{rp.object}{A ResponsePatterns object.}

\item{percentile}{An integer. Defines a percentile cutoff.}
}
\value{
A ResponsePatterns object.
}
\description{
This function reorders observations and selects those equal of above a defined percentile.
}
\examples{
rp <- rp.acors(rp.simdata, id.var="optional_ID")
rp <- rp.select(rp, percentile=80)
}
\seealso{
\code{\link{rp.acors}}, \code{\link{rp.patterns}}
}
