% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.summary.R
\name{rp.summary}
\alias{rp.summary}
\title{SUmmary of an ResponsePatterns object}
\usage{
rp.summary(rp.object)
}
\arguments{
\item{rp.object}{A ResponsePatterns object.}
}
\value{
Prints a summary of a ResponsePatterns object.
}
\description{
SUmmary of an ResponsePatterns object
}
\examples{
rp <- rp.acors(rp.simdata, id.var="optional_ID")
rp.summary(rp)
summary(rp)
}
