% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_select.R
\name{rp.select}
\alias{rp.select}
\title{Select observations}
\usage{
rp.select(rp.object, percentile = 90)
}
\arguments{
\item{rp.object}{A ResponsePatterns object.}

\item{percentile}{An integer. Defines a percentile cutoff.}
}
\value{
A ResponsePatterns object.
}
\description{
This function reorders observations and selects those equal of above a defined percentile.
}
\examples{
rp.object <- rp.select(rp.object, percentile=80)
rp.acors(data) \%>\% rp.select(percentile=90) \%>\% rp.indices()
}
